End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import {MessageService} from '../_services/message.service';
 
 
@@ -17,7 +18,7 @@ pageSize =5;
 pageNumber =1;
 loading = false;
 
-  constructor(private messageService: MessageService ) { 
+  constructor(private messageService: MessageService, private confirmService: ConfirmService ) { 
 
 
   }
@@ -39,9 +40,14 @@ loading = false;
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-this.messages.splice(this.messages.findIndex(m=>m.id === id),1);
-    });
+    this.confirmService.confirm('Confirm delete message','This cannot be undone').subscribe(result =>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m=>m.id === id),1);
+              });
+      }
+    })
+
   }
